Applied CR suggestion to design code differently
