+++ .vib/spring-cloud-dataflow/cypress/cypress/integration/spring-cloud-dataflow_spec.js
@@ -1,10 +1,10 @@
 /// <reference types="cypress" />
+import { random } from './utils';
 import {
-  random,
   importAStreamApplication,
   importATaskApplication,
   createATask,
-} from './utils';
+} from '/prepare-app-state';
 
 it('allows getting Spring Cloud Dataflow info', () => {
   cy.visit('/dashboard');
@@ -16,6 +16,7 @@ it('allows getting Spring Cloud Dataflow info', () => {
     .and('contain', 'Shell');
 });
 
+//this is a comment
 it('allows a stream to be created', () => {
   const STREAM = 'mongodb | cassandra';
